Move function for finding last event to base interactor
