+++ opensrp-reveal/src/main/java/org/smartregister/reveal/contract/BaseContract.java
@@ -6,6 +6,7 @@ import com.mapbox.geojson.Feature;
 
 import org.json.JSONArray;
 import org.smartregister.commonregistry.CommonPersonObjectClient;
+import org.smartregister.domain.Event;
 import org.smartregister.domain.Task;
 
 /**
@@ -28,5 +29,9 @@ public interface BaseContract {
     interface BaseInteractor {
 
         void saveJsonForm(String json);
+
+        void handleLasteventFound(Event event);
+
+        void findLastEvent(String eventBaseEntityId, String eventType);
     }
 }
